update test system

+++ src/main/java/com/example/demo/controller/UserDeckController.java
@@ -20,7 +20,7 @@ public class UserDeckController {
     }
 
     @GetMapping("/{id}")
-    public UserDeck getUserDeckById(@PathVariable int id) {
+    public UserDeck getUserDeckById(@PathVariable Long id) {
         return userDeckService.getUserDeckById(id);
     }
